Fix some styling
