+++ frontend/app/screens/CampusMap.tsx
@@ -1,13 +1,12 @@
 //Campus Map Screen
 import { StyleSheet, Text, View, TextInput, Image } from "react-native";
-
 import OutdoorMap from "../components/maps/OutdoorMap";
-import { StatusBar } from "expo-status-bar";
 import AppLogo from "../components/ui/AppLogo";
 import { ThemeProvider } from "../components/context/ThemeContext";
 import ThemeToggle from "../components/ui/ThemeToggle";
 import useTheme from "../hooks/useTheme";
 import { getStyles } from "../styles";
+import CampusPilotHeader from "../components/ui/CampusPilotHeader";
 
 export default function CampusMap() {
   const { theme, toggleTheme } = useTheme(); // Use the custom hook
@@ -15,32 +14,18 @@ export default function CampusMap() {
 
   return (
     <View style={styles.container}>
-      {/* Adding the status bar on top of the screen */}
-      <View>
-        <StatusBar style="dark" />
-      </View>
-      <View style={styles.iconRow}>
-        <View style={styles.iconRowView}>
-          <AppLogo />
-        </View>
-        <View>
-          <Text style={(styles.iconRowView, styles.logo)}>Campus Pilot</Text>
-        </View>
-        {/*Dark/light theme toggle*/}
-        <ThemeToggle />
-      </View>
+      {/* Header with logo , title and Dakr/Light theme toggle */}
+      <CampusPilotHeader />
 
-      {/*SearchBar*/}
+      {/*SearchBar
       <View style={styles.searchBar}>
         <TextInput placeholder={"Search"} placeholderTextColor={"#666"} />
-      </View>
+      </View>*/}
+
       {/*Campus Map*/}
       <View style={styles.mapContainer}>
         <OutdoorMap />
       </View>
     </View>
   );
-}
-{
-  /*Styles for the views*/
 }
